requirements.txt
# Fury AI Assistant — Requirements
# Install with: pip install -r requirements.txt

# Core
openai>=1.0.0
python-dotenv>=1.0.0
groq>=0.4.0

# Automation
pyautogui>=0.9.54
pynput>=1.7.6
pywinauto>=0.6.8
pywin32>=306

# Vision
mss>=9.0.1
opencv-python>=4.8.0
pytesseract>=0.3.10
Pillow>=10.0.0
numpy>=1.24.0

# Browser
playwright>=1.40.0

# Voice
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# Web / API
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data
sqlite3-api>=2.0.1

# Utilities
schedule>=1.2.0
psutil>=5.9.0
watchdog>=3.0.0
pyyaml>=6.0.1